fix(core): replace regex with balanced parser to support nested JSON
