+++ core/framework/graph/node.py
@@ -24,47 +24,12 @@ from pydantic import BaseModel, Field
 
 from framework.runtime.core import Runtime
 from framework.llm.provider import LLMProvider, Tool
+from framework.utils import find_json_object
 
 logger = logging.getLogger(__name__)
 
 
-def find_json_object(text: str) -> str | None:
-    """Find the first valid JSON object in text using balanced brace matching.
 
-    This handles nested objects correctly, unlike simple regex like r'\\{[^{}]*\\}'.
-    """
-    start = text.find('{')
-    if start == -1:
-        return None
-
-    depth = 0
-    in_string = False
-    escape_next = False
-
-    for i, char in enumerate(text[start:], start):
-        if escape_next:
-            escape_next = False
-            continue
-
-        if char == '\\' and in_string:
-            escape_next = True
-            continue
-
-        if char == '"' and not escape_next:
-            in_string = not in_string
-            continue
-
-        if in_string:
-            continue
-
-        if char == '{':
-            depth += 1
-        elif char == '}':
-            depth -= 1
-            if depth == 0:
-                return text[start:i + 1]
-
-    return None
 
 
 class NodeSpec(BaseModel):
